+++ NEWS.md
@@ -6,6 +6,8 @@
 - Fixed a regression which caused calls to an overloaded subprogram with
   the same name as an enclosing subprogram to be incorrectly reported as
   ambiguous (#1560).
+- Fixed reading scalar `out` and `inout` parameters in VHPI foreign
+  subprograms (#1570).
 - Several other minor bugs were resolved (#1559, #1562).
 
 ## Version 1.21.0 - 2026-05-23